Sprint task - ALM: allow central definition of repository (project) and pass fail option.

+++ org.eclipse.jubula.tools/src/org/eclipse/jubula/tools/jarutils/IVersion.java
@@ -24,7 +24,7 @@ public interface IVersion {
     public final Integer JB_PROTOCOL_MAJOR_VERSION = new Integer(11);
     
     /** major version for DB */
-    public final Integer JB_DB_MAJOR_VERSION = new Integer(42);
+    public final Integer JB_DB_MAJOR_VERSION = new Integer(43);
     
     /** minor version for DB */
     public final Integer JB_DB_MINOR_VERSION = new Integer(0);
